Consolidated fan logic into a shared module
